+++ src/imio/annex/__init__.py
@@ -3,5 +3,8 @@
 
 from zope.i18nmessageid import MessageFactory
 
+import logging
+
 
 _ = MessageFactory('imio.annex')
+logger = logging.getLogger('imio.annex')