+++ src/imio/annex/events.py
@@ -8,8 +8,8 @@ Created by mpeeters
 """
 
 from imio.annex.interfaces import IAnnexFileChangedEvent
-from imio.annex.interfaces import IConversionStartedEvent
 from imio.annex.interfaces import IConversionReallyFinishedEvent
+from imio.annex.interfaces import IConversionStartedEvent
 from zope.component.interfaces import ObjectEvent
 from zope.interface import implements